Usuarios y oficina

+++ SIMEXPRO Backend/SIMEXPRO_API/SIMEXPRO.BusinessLogic/Services/GeneralServices/GeneralServices.cs
@@ -694,26 +694,40 @@ namespace SIMEXPRO.BussinessLogic.Services.GeneralServices
         #region Oficinas
         public ServiceResult ListarOficinas()
         {
-            var result = new ServiceResult();
+            var resultado = new ServiceResult();
             try
             {
                 var list = _oficinasRepository.List();
-                return result.Ok(list);
+                return resultado.Ok(list);
             }
             catch (Exception ex)
             {
-                return result.Error(ex.Message);
+                return resultado.Error(ex.Message);
             }
         }
 
         public ServiceResult InsertarOficinas(tbOficinas item)
         {
-
             var result = new ServiceResult();
             try
             {
-                var list = _oficinasRepository.Insert(item);
-                return result.Ok(list);
+                if (item.ofic_Nombre != "")
+                {
+                    var map = _oficinasRepository.Insert(item);
+                    if (map.MessageStatus == "1")
+                    {
+                        return result.Ok(map);
+                    }
+                    else
+                    {
+                        map.MessageStatus = (map.CodeStatus == 0) ? "401 Error de Consulta" : map.MessageStatus;
+                        return result.Error(map);
+                    }
+                }
+                else
+                {
+                    return result.SetMessage("La solicitud contiene sintaxis erronea", ServiceResultType.BadRecuest);
+                }
             }
             catch (Exception ex)
             {
@@ -726,8 +740,23 @@ namespace SIMEXPRO.BussinessLogic.Services.GeneralServices
             var result = new ServiceResult();
             try
             {
-                var list = _oficinasRepository.Update(item);
-                return result.Ok(list);
+                if (item.ofic_Nombre != "")
+                {
+                    var map = _oficinasRepository.Update(item);
+                    if (map.MessageStatus == "1")
+                    {
+                        return result.Ok(map);
+                    }
+                    else
+                    {
+                        map.MessageStatus = (map.CodeStatus == 0) ? "401 Error de Consulta" : map.MessageStatus;
+                        return result.Error(map);
+                    }
+                }
+                else
+                {
+                    return result.SetMessage("La solicitud contiene sintaxis erronea", ServiceResultType.BadRecuest);
+                }
             }
             catch (Exception ex)
             {
@@ -737,18 +766,31 @@ namespace SIMEXPRO.BussinessLogic.Services.GeneralServices
 
         public ServiceResult EliminarOficinas(tbOficinas item)
         {
-
             var result = new ServiceResult();
             try
             {
-                var list = _oficinasRepository.Delete(item);
-                return result.Ok(list);
+                if (item.ofic_Id != 0)
+                {
+                    var map = _oficinasRepository.Delete(item);
+                    if (map.CodeStatus > 0)
+                    {
+                        return result.Ok(map);
+                    }
+                    else
+                    {
+                        map.MessageStatus = (map.CodeStatus == 0) ? "401 Error de Consulta" : map.MessageStatus;
+                        return result.Error(map);
+                    }
+                }
+                else
+                {
+                    return result.SetMessage("La solicitud contiene sintaxis erronea", ServiceResultType.BadRecuest);
+                }
             }
             catch (Exception ex)
             {
                 return result.Error(ex.Message);
             }
-
         }
         #endregion
